+++ src/apps/sequencer/model/Settings.h
@@ -6,7 +6,7 @@
 
 class Settings {
 public:
-    static constexpr uint32_t Version = 4;
+    static constexpr uint32_t Version = 5;
 
     static const char *Filename;